newscast: diffuse our ID with a higher timestamp during the bootstrap period
In typical applications, peers bootstrap from the same
small set of nodes. Therefore, at the beginning, their
cache is not a uniform random sample. As a consequence,
their messages could hit bootstrap nodes aggressively,
also modifying the randomness of those caches.

We avoid this by implementing an initial grace period
when the entering peer's cache is randomized without
disturbing the system as a whole.